move word correction into audio toolkit
