+++ src-tauri/src/audio_toolkit/mod.rs
@@ -1,8 +1,10 @@
 pub mod audio;
 pub mod constants;
+pub mod text;
 pub mod vad;
 
 pub use audio::{
     list_input_devices, list_output_devices, save_wav_file, AudioRecorder, CpalDeviceInfo,
 };
+pub use text::apply_custom_words;
 pub use vad::{SileroVad, VoiceActivityDetector};
