+++ src-tauri/src/managers/transcription.rs
@@ -1,14 +1,13 @@
+use crate::audio_toolkit::apply_custom_words;
 use crate::managers::model::{EngineType, ModelManager};
 use crate::settings::{get_settings, ModelUnloadTimeout};
 use anyhow::Result;
 use log::debug;
-use natural::phonetics::soundex;
 use serde::Serialize;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::{Duration, SystemTime};
-use strsim::levenshtein;
 use tauri::{App, AppHandle, Emitter, Manager};
 use transcribe_rs::{
     engines::{
@@ -42,114 +41,6 @@ pub struct TranscriptionManager {
     last_activity: Arc<AtomicU64>,
     shutdown_signal: Arc<AtomicBool>,
     watcher_handle: Arc<Mutex<Option<thread::JoinHandle<()>>>>,
-}
-
-fn apply_custom_words(text: &str, custom_words: &[String], threshold: f64) -> String {
-    if custom_words.is_empty() {
-        return text.to_string();
-    }
-
-    // Pre-compute lowercase versions to avoid repeated allocations
-    let custom_words_lower: Vec<String> = custom_words.iter().map(|w| w.to_lowercase()).collect();
-
-    let words: Vec<&str> = text.split_whitespace().collect();
-    let mut corrected_words = Vec::new();
-
-    for word in words {
-        let cleaned_word = word
-            .trim_matches(|c: char| !c.is_alphabetic())
-            .to_lowercase();
-
-        if cleaned_word.is_empty() {
-            corrected_words.push(word.to_string());
-            continue;
-        }
-
-        // Skip extremely long words to avoid performance issues
-        if cleaned_word.len() > 50 {
-            corrected_words.push(word.to_string());
-            continue;
-        }
-
-        let mut best_match: Option<&String> = None;
-        let mut best_score = f64::MAX;
-
-        for (i, custom_word_lower) in custom_words_lower.iter().enumerate() {
-            // Skip if lengths are too different (optimization)
-            let len_diff = (cleaned_word.len() as i32 - custom_word_lower.len() as i32).abs();
-            if len_diff > 5 {
-                continue;
-            }
-
-            // Calculate Levenshtein distance (normalized by length)
-            let levenshtein_dist = levenshtein(&cleaned_word, custom_word_lower);
-            let max_len = cleaned_word.len().max(custom_word_lower.len()) as f64;
-            let levenshtein_score = if max_len > 0.0 {
-                levenshtein_dist as f64 / max_len
-            } else {
-                1.0
-            };
-
-            // Calculate phonetic similarity using Soundex
-            let phonetic_match = soundex(&cleaned_word, custom_word_lower);
-
-            // Combine scores: favor phonetic matches, but also consider string similarity
-            let combined_score = if phonetic_match {
-                levenshtein_score * 0.3 // Give significant boost to phonetic matches
-            } else {
-                levenshtein_score
-            };
-
-            // Accept if the score is good enough (configurable threshold)
-            if combined_score < threshold && combined_score < best_score {
-                best_match = Some(&custom_words[i]);
-                best_score = combined_score;
-            }
-        }
-
-        if let Some(replacement) = best_match {
-            // Preserve the original case pattern as much as possible
-            let corrected = if word.chars().all(|c| c.is_uppercase()) {
-                replacement.to_uppercase()
-            } else if word.chars().next().map_or(false, |c| c.is_uppercase()) {
-                let mut chars: Vec<char> = replacement.chars().collect();
-                if let Some(first_char) = chars.get_mut(0) {
-                    *first_char = first_char.to_uppercase().next().unwrap_or(*first_char);
-                }
-                chars.into_iter().collect()
-            } else {
-                replacement.clone()
-            };
-
-            // Preserve punctuation from original word - optimized version
-            let prefix_end = word.chars().take_while(|c| !c.is_alphabetic()).count();
-            let suffix_start = word
-                .char_indices()
-                .rev()
-                .take_while(|(_, c)| !c.is_alphabetic())
-                .count();
-
-            let original_prefix = if prefix_end > 0 {
-                &word[..prefix_end]
-            } else {
-                ""
-            };
-            let original_suffix = if suffix_start > 0 {
-                &word[word.len() - suffix_start..]
-            } else {
-                ""
-            };
-
-            corrected_words.push(format!(
-                "{}{}{}",
-                original_prefix, corrected, original_suffix
-            ));
-        } else {
-            corrected_words.push(word.to_string());
-        }
-    }
-
-    corrected_words.join(" ")
 }
 
 impl TranscriptionManager {
